MCMC Routine for Bayesian Dynamic Factor Models

CppBDFM is the core C++ function for estimating a linear-Gaussian Bayesain dynamic factor model by MCMC methods using Durbin and Koopman's disturbance smoother. This function may be called directly by advanced users. The only dependencies are the Armadillo (http://arma.sourceforge.net/) linear algebra library for C++ and the packages needed for interfacing with R (Rcpp and RcppArmadillo).

CppBDFM(B, Bp, lam_B, q, nu_q, H, Hp, lam_H, R, nu_r, Y, reps, burn, Loud)

Arguments

B
initial guess for B in transition equation

Bp
prior for B

lam_B
prior tightness for B (additive)

q
initial guess for q in the transition equation

nu_q
prior "degrees of freedom" for inverse-Whishart prior for q (additive, prior scale is fixed so that increasing nu_q shrinks the variance towards zero)

H
initial guess for H in the trasition equation

Hp
prior for H

lam_H
prior tightness for H (additive)

R
initial guess for diagonal elements of R in the transition equation, entered as a vector

nu_r
prior deg. of freedom for elements of R, entered as a vector (additive, prior scale is fixed so that increasing nu_r[j] shrinks the variance of shocks to series j towards zero)

Y
Input data. Data must be scaled and centered prior to estimation if desired.

reps
number of repetitions for MCMC sampling

burn
number of iterations to burn in MCMC sampling

Loud
print status of function during evaluation.