Estimate dynamic factor model

Estimate dynamic factor model

dfm(Y, factors = 1, lags = 2, forecast = 0, method = c("bayesian",
  "ml", "pc"), B_prior = NULL, lam_B = 0, H_prior = NULL,
  lam_H = 0, nu_q = 0, nu_r = NULL, identification = "PC_full",
  intercept = TRUE, reps = 1000, burn = 500, loud = FALSE,
  EM_tolerance = 0.01)

Arguments

Y
data in matrix format with time in rows

factors
number of factors

lags
number of lags in transition equation

forecast
number of periods ahead to forecast

method
character, method to be used

B_prior
prior matrix for B in the transition equation. Default is zeros.

lam_B
prior tightness on B

H_prior
prior matrix for H (loadings) in the observation equation. Default is zeros.

lam_H
prior tightness on H

nu_q
prior deg. of freedom for transition equation, entered as vector with length equal to the number of factors.

nu_r
prior deg. of freedom for observables, entered as vector with length equal to the number of observables.

identification
factor identification. 'PC_full' is the default (using all observed series), 'PC_sub' finds a submatrix of the data that maximizes the number of observations for a square (no missing values) data set. Use 'PC_sub' when many observations are missing.

intercept
logical, should an icept be included?

reps
number of repetitions for MCMC sampling

burn
number of iterations to burn in MCMC sampling

loud
print status of function during evalutation. If ML, print difference in likelihood at each iteration of the EM algorithm.

EM_tolerance
tolerance for convergence of EM algorithm (method `ml` only). Convergence criteria is calculated as 200 * (Lik1 - Lik0) / abs(Lik1 + Lik0) where Lik1 is the log likelihood from this iteration and Lik0 is the likelihood from the previous iteration.

Examples

fdeaths0 <- fdeaths
fdeaths0[length(fdeaths0)] <- NA
dta <- cbind(fdeaths0, mdeaths)

library(BDFM)
m <- dfm(dta, forecast = 2)

#> Loading required namespace: tsbox

predict(m)

#>           fdeaths0   mdeaths
#> Jan 1974  820.0496 2136.9894
#> Feb 1974  709.1714 1863.1459
#> Mar 1974  717.7432 1884.3162
#> Apr 1974  713.3606 1873.4921
#> May 1974  558.0762 1489.9757
#> Jun 1974  459.3264 1246.0864
#> Jul 1974  471.8216 1276.9468
#> Aug 1974  412.4125 1130.2198
#> Sep 1974  442.9569 1205.6575
#> Oct 1974  559.2180 1492.7957
#> Nov 1974  610.3722 1619.1347
#> Dec 1974  701.3139 1843.7395
#> Jan 1975  806.6398 2103.8704
#> Feb 1975  818.4413 2133.0172
#> Mar 1975  825.1282 2149.5324
#> Apr 1975  695.9599 1830.5165
#> May 1975  521.6876 1400.1042
#> Jun 1975  475.1213 1285.0961
#> Jul 1975  434.7101 1185.2899
#> Aug 1975  413.3003 1132.4125
#> Sep 1975  380.5507 1051.5286
#> Oct 1975  498.1589 1341.9936
#> Nov 1975  579.6323 1543.2141
#> Dec 1975  789.6577 2061.9283
#> Jan 1976  774.0148 2023.2941
#> Feb 1976 1068.8315 2751.4233
#> Mar 1976  879.6000 2284.0651
#> Apr 1976  553.5180 1478.7179
#> May 1976  436.4870 1189.6784
#> Jun 1976  424.2923 1159.5604
#> Jul 1976  404.5483 1110.7973
#> Aug 1976  347.2354  969.2475
#> Sep 1976  359.3140  999.0788
#> Oct 1976  443.7452 1207.6044
#> Nov 1976  549.1598 1467.9543
#> Dec 1976  787.6513 2056.9731
#> Jan 1977  860.8862 2237.8464
#> Feb 1977  618.5114 1639.2367
#> Mar 1977  652.2441 1722.5486
#> Apr 1977  682.1323 1796.3655
#> May 1977  461.1541 1250.6005
#> Jun 1977  424.2128 1159.3639
#> Jul 1977  395.3681 1088.1241
#> Aug 1977  364.4484 1011.7598
#> Sep 1977  344.6136  962.7724
#> Oct 1977  430.4775 1174.8364
#> Nov 1977  451.7606 1227.4006
#> Dec 1977  625.0287 1655.3327
#> Jan 1978  773.1065 2021.0508
#> Feb 1978  878.4212 2281.1538
#> Mar 1978  740.8756 1941.4479
#> Apr 1978  531.8519 1425.2076
#> May 1978  497.9233 1341.4117
#> Jun 1978  435.8214 1188.0344
#> Jul 1978  400.1892 1100.0312
#> Aug 1978  361.3433 1004.0909
#> Sep 1978  348.8920  973.3390
#> Oct 1978  416.1384 1139.4222
#> Nov 1978  405.6554 1113.5316
#> Dec 1978  688.0446 1810.9676
#> Jan 1979  868.8154 2257.4295
#> Feb 1979  695.1554 1828.5296
#> Mar 1979  702.3399 1846.2736
#> Apr 1979  575.7578 1533.6450
#> May 1979  447.6945 1217.3583
#> Jun 1979  391.1854 1077.7940
#> Jul 1979  382.8878 1057.3006
#> Aug 1979  350.3998  977.0629
#> Sep 1979  337.0814  944.1696
#> Oct 1979  393.0897 1082.4971
#> Nov 1979  478.7870 1294.1496
#> Dec 1979  497.8981 1341.3497
#> Jan 1980  528.3557 1416.5727
#> Feb 1980  553.6721 1479.0986