Dates of Chinese New Year, Indian Diwali and Easter

Dates of Chinese New Year, Indian Diwali and Easter, suitable for the use in [seas_we()].

Format

Objects of class `"Date"`.

Source

http://www.census.gov/srd/www/genhol/genhol_downloads.html

http://www.chinesenewyears.info/chinese-new-year-calendar.php

Ministry of Statistics and Programme Implementation

Examples

cny

#>   [1] "1930-01-29" "1931-02-17" "1932-02-06" "1933-01-26" "1934-02-14"
#>   [6] "1935-02-04" "1936-01-24" "1937-02-11" "1938-01-31" "1939-02-19"
#>  [11] "1940-02-08" "1941-01-27" "1942-02-15" "1943-02-04" "1944-01-25"
#>  [16] "1945-02-13" "1946-02-01" "1947-01-22" "1948-02-10" "1949-01-29"
#>  [21] "1950-02-17" "1951-02-06" "1952-01-27" "1953-02-14" "1954-02-03"
#>  [26] "1955-01-24" "1956-02-12" "1957-01-31" "1958-02-18" "1959-02-08"
#>  [31] "1960-01-28" "1961-02-15" "1962-02-05" "1963-01-25" "1964-02-13"
#>  [36] "1965-02-02" "1966-01-21" "1967-02-09" "1968-01-30" "1969-02-17"
#>  [41] "1970-02-06" "1971-01-27" "1972-02-15" "1973-02-03" "1974-01-23"
#>  [46] "1975-02-11" "1976-01-31" "1977-02-18" "1978-02-07" "1979-01-28"
#>  [51] "1980-02-16" "1981-02-05" "1982-01-25" "1983-02-13" "1984-02-02"
#>  [56] "1985-02-20" "1986-02-09" "1987-01-29" "1988-02-17" "1989-02-06"
#>  [61] "1990-01-27" "1991-02-15" "1992-02-04" "1993-01-23" "1994-02-10"
#>  [66] "1995-01-31" "1996-02-19" "1997-02-07" "1998-01-28" "1999-02-16"
#>  [71] "2000-02-05" "2001-01-24" "2002-02-12" "2003-02-01" "2004-01-22"
#>  [76] "2005-02-09" "2006-01-29" "2007-02-18" "2008-02-07" "2009-01-26"
#>  [81] "2010-02-14" "2011-02-03" "2012-01-23" "2013-02-10" "2014-01-31"
#>  [86] "2015-02-19" "2016-02-08" "2017-01-28" "2018-02-16" "2019-02-05"
#>  [91] "2020-01-25" "2021-02-12" "2022-02-01" "2023-01-22" "2024-02-10"
#>  [96] "2025-01-29" "2026-02-17" "2027-02-06" "2028-01-26" "2029-02-13"
#> [101] "2030-02-03"

diwali

#>   [1] "1900-10-23" "1901-11-11" "1902-10-31" "1903-10-20" "1904-11-07"
#>   [6] "1905-10-28" "1906-10-17" "1907-11-05" "1908-10-25" "1909-11-13"
#>  [11] "1910-11-02" "1911-10-22" "1912-11-09" "1913-10-29" "1914-10-19"
#>  [16] "1915-11-07" "1916-10-26" "1917-11-14" "1918-11-03" "1919-10-23"
#>  [21] "1920-11-10" "1921-10-30" "1922-10-20" "1923-11-08" "1924-10-28"
#>  [26] "1925-10-17" "1926-11-05" "1927-10-25" "1928-11-12" "1929-11-01"
#>  [31] "1930-10-21" "1931-11-09" "1932-10-29" "1933-10-19" "1934-11-07"
#>  [36] "1935-10-27" "1936-11-14" "1937-11-03" "1938-10-23" "1939-11-11"
#>  [41] "1940-10-30" "1941-10-20" "1942-11-08" "1943-10-29" "1944-10-17"
#>  [46] "1945-11-04" "1946-10-24" "1947-11-12" "1948-11-01" "1949-10-21"
#>  [51] "1950-11-09" "1951-10-30" "1952-10-18" "1953-11-06" "1954-10-26"
#>  [56] "1955-11-14" "1956-11-02" "1957-10-23" "1958-11-11" "1959-10-31"
#>  [61] "1960-10-20" "1961-11-08" "1962-10-28" "1963-10-17" "1964-11-04"
#>  [66] "1965-10-24" "1966-11-12" "1967-11-02" "1968-10-21" "1969-11-09"
#>  [71] "1970-10-30" "1971-10-19" "1972-11-06" "1973-10-26" "1974-11-14"
#>  [76] "1975-11-03" "1976-10-23" "1977-11-11" "1978-10-31" "1979-10-21"
#>  [81] "1980-11-07" "1981-10-27" "1982-11-15" "1983-11-04" "1984-10-24"
#>  [86] "1985-11-12" "1986-11-01" "1987-10-22" "1988-11-09" "1989-10-29"
#>  [91] "1990-10-18" "1991-11-06" "1992-10-25" "1993-11-13" "1994-11-03"
#>  [96] "1995-10-23" "1996-11-10" "1997-10-30" "1998-10-19" "1999-11-07"
#> [101] "2000-10-26" "2001-11-14" "2002-11-04" "2003-10-25" "2004-11-12"
#> [106] "2005-11-01" "2006-10-21" "2007-11-09" "2008-10-28" "2009-10-17"
#> [111] "2010-11-05" "2011-10-26" "2012-11-13" "2013-11-02" "2014-10-22"
#> [116] "2015-11-11" "2016-10-30" "2017-10-19" "2018-11-07" "2019-10-27"
#> [121] "2020-11-14" "2021-11-04" "2022-10-24" "2023-11-12" "2024-11-01"
#> [126] "2025-10-21" "2026-11-08" "2027-10-29" "2028-10-17" "2029-11-05"
#> [131] "2030-10-26"

easter

#>   [1] "1931-04-05" "1932-03-27" "1933-04-16" "1934-04-01" "1935-04-21"
#>   [6] "1936-04-12" "1937-03-28" "1938-04-17" "1939-04-09" "1940-03-24"
#>  [11] "1941-04-13" "1942-04-05" "1943-04-25" "1944-04-09" "1945-04-01"
#>  [16] "1946-04-21" "1947-04-06" "1948-03-28" "1949-04-17" "1950-04-09"
#>  [21] "1951-03-25" "1952-04-13" "1953-04-05" "1954-04-18" "1955-04-10"
#>  [26] "1956-04-01" "1957-04-21" "1958-04-06" "1959-03-29" "1960-04-17"
#>  [31] "1961-04-02" "1962-04-22" "1963-04-14" "1964-03-29" "1965-04-18"
#>  [36] "1966-04-10" "1967-03-26" "1968-04-14" "1969-04-06" "1970-03-29"
#>  [41] "1971-04-11" "1972-04-02" "1973-04-22" "1974-04-14" "1975-03-30"
#>  [46] "1976-04-18" "1977-04-10" "1978-03-26" "1979-04-15" "1980-04-06"
#>  [51] "1981-04-19" "1982-04-11" "1983-04-03" "1984-04-22" "1985-04-07"
#>  [56] "1986-03-30" "1987-04-19" "1988-04-03" "1989-03-26" "1990-04-15"
#>  [61] "1991-03-31" "1992-04-19" "1993-04-11" "1994-04-03" "1995-04-16"
#>  [66] "1996-04-07" "1997-03-30" "1998-04-12" "1999-04-04" "2000-04-23"
#>  [71] "2001-04-15" "2002-03-31" "2003-04-20" "2004-04-11" "2005-03-27"
#>  [76] "2006-04-16" "2007-04-08" "2008-03-23" "2009-04-12" "2010-04-04"
#>  [81] "2011-04-24" "2012-04-08" "2013-03-31" "2014-04-20" "2015-04-05"
#>  [86] "2016-03-27" "2017-04-16" "2018-04-01" "2019-04-21" "2020-04-12"
#>  [91] "2021-04-04" "2022-04-17" "2023-04-09" "2024-03-31" "2025-04-20"
#>  [96] "2026-04-05" "2027-03-28" "2028-04-16" "2029-04-01" "2030-04-21"